update :: token parser 분리
