+++ src/main/kotlin/com/msg/gauth/global/filter/JwtTokenFilter.kt
@@ -1,6 +1,7 @@
 package com.msg.gauth.global.filter
 
 import com.msg.gauth.global.security.jwt.JwtTokenProvider
+import com.msg.gauth.global.security.jwt.TokenParser
 import org.springframework.security.core.context.SecurityContextHolder
 import org.springframework.web.filter.OncePerRequestFilter
 import javax.servlet.FilterChain
@@ -8,6 +9,7 @@ import javax.servlet.http.HttpServletRequest
 import javax.servlet.http.HttpServletResponse
 
 class JwtTokenFilter(
+    private val tokenParser: TokenParser,
     private val jwtTokenProvider: JwtTokenProvider
 ): OncePerRequestFilter() {
 
@@ -18,7 +20,7 @@ class JwtTokenFilter(
     ) {
         val token: String? = jwtTokenProvider.resolveToken(request)
         if (!token.isNullOrBlank()) {
-            val authentication = jwtTokenProvider.authentication(token)
+            val authentication = tokenParser.authentication(token)
             SecurityContextHolder.getContext().authentication = authentication
         }
         filterChain.doFilter(request, response)
